economics: make integer Curve the authoritative sim default
shekyl-staking-sim now defaults to --curve-impl=integer (canonical
reward_arithmetic, the minting code). Float becomes opt-in exploration
only: it over-reads worst-shard redundancy by up to one replica (PR 1.5
tail-margin finding), so it is no longer a co-authoritative validation
substrate. Unknown --curve-impl values fail loudly (exit 2) instead of
silently falling back to float; each run prints the active backend.
CurveImpl type default and scenario base config flip to Integer; the
float<->integer reconciliation tests remain as drift guards.

ARCHIVAL_SIM_ECONOMICS_VERDICT.md; REWARD_EMISSION_VIN_PLAN.md M-1.

+++ rust/shekyl-staking-sim/src/scenarios.rs
@@ -1701,7 +1701,9 @@ fn baseline() -> SimConfig {
         cap: 8.0,
         pseudonym_cost: 0.05,
         age_weight: 2.0,
-        curve_impl: CurveImpl::Float,
+        // Authoritative backend by default (the canonical minting arithmetic);
+        // main overrides per `--curve-impl`. Float is opt-in exploration only.
+        curve_impl: CurveImpl::Integer,
         storage_unit_cost: 0.03,
         bond_rate: 2.0,      // mid
         bond_age_scale: 0.0, // flat bond is the iteration-1 baseline (L4 sweep varies it)
